requirements.txt
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit 
phidata
google-generativeai
openai
pandas